# Fix qoibench handling of RGB input