There's a flaw in the way qoibench handles RGB which makes libpng/stbi perform much worse than they should. The way it currently works is that all input is converted to RGBA, so the encoders encode to RGBA always. QOI doesn't really care aside from reading 3-4 bytes per pixel, but for PNG/STBI it's very important as RGB is encoded as RGBA with all the overhead that implies. Treating all input as RGBA also juices the rate stat across the board.

This pull request:
* Reads the input image to have the encoders work from RGB input if it has 3 channels, RGBA otherwise
* Calculates the rate relative to `w*h*channels`, instead of always `w*h*4` (channels will now be 3 when input is RGB)

It's not a perfect solution as all non-RGB input is treated as RGBA still, but it fixes the bulk of the issue. Handling paletted, gray and gray + alpha properly is probably more hassle than it's worth, dealing with stbi is a bit of a pain.

Before
```
# Grand total for ../runtime/qoi_benchmark_suite/images
        decode ms   encode ms   decode mpps   encode mpps   size kb    rate
libpng:       8.4       117.5         54.99          3.95       423   23.3%
stbi:         8.5        78.8         54.45          5.89       601   33.2%
qoi:          2.3         3.2        200.35        144.85       463   25.6%
```

After
```
# Grand total for ../runtime/qoi_benchmark_suite/images
        decode ms   encode ms   decode mpps   encode mpps   size kb    rate
libpng:       9.0       108.4         51.69          4.28       398   24.2%
stbi:         9.1        76.7         51.14          6.05       561   34.2%
qoi:          2.5         3.6        187.50        130.17       463   28.2%
```

PS hopefully this is the correct way to do a pull request.