@phoboslab 2021-12-14
Fair point! Although I have to say that libpng and stbi _should_ maybe perform the same, no matter how many channels you specify. Just like QOI does :D